Fix Julia 0.2 compatibility

+++ src/binomial.jl
@@ -64,8 +64,8 @@ immutable SignTest <: HypothesisTest
 end
 
 SignTest{T<:Real}(x::AbstractVector{T}, median::Real=0) =
-	SignTest(median, sum(x .> median), sum(x .!= median))
-SignTest{T<:Real, S<:Real}(x::AbstractVector{T}, y::AbstractVector{S}) = SignTest(x - y, 0)
+	SignTest(float64(median), sum(x .> median), sum(x .!= median))
+SignTest{T<:Real, S<:Real}(x::AbstractVector{T}, y::AbstractVector{S}) = SignTest(x - y, 0.0)
 
 pvalue(x::SignTest; tail=:both) = pvalue(Binomial(x.n, 0.5), x.x; tail=tail)
